refactor: Rename modules: `data` to `serde` and `json_model` to `json`.
