+++ src/model/_impl_xml_serde.rs
@@ -1,4 +1,4 @@
-use crate::data::xml_model::*;
+use crate::serde::xml_model::*;
 use crate::update_fn::bma_fn_update::BmaFnUpdate;
 use crate::{
     BmaLayout, BmaLayoutContainer, BmaLayoutVariable, BmaModel, BmaNetwork, BmaRelationship,
