+++ src/model/bma_network.rs
@@ -20,7 +20,7 @@ pub struct BmaNetwork {
 }
 
 impl BmaNetwork {
-    /// Create a new [`BmaNetwork`] from the provided data.
+    /// Create a new [`BmaNetwork`] from the provided serde.
     pub fn new(variables: Vec<BmaVariable>, relationships: Vec<BmaRelationship>) -> Self {
         BmaNetwork {
             name: None,
