+++ src/serde/json/json_layout_variable.rs
@@ -1,4 +1,4 @@
-use crate::data::quote_num::QuoteNum;
+use crate::serde::quote_num::QuoteNum;
 use crate::utils::{f64_or_default, rational_or_default, take_if_not_blank};
 use crate::{BmaLayoutVariable, VariableType};
 use serde::{Deserialize, Serialize};
