+++ src/serde/mod.rs
@@ -1,4 +1,4 @@
-pub(crate) mod json_model;
+pub(crate) mod json;
 pub(crate) mod xml_model;
 
 pub(crate) mod quote_num;
